fix pr comment
